+++ app/database/migrations/2024_08_17_000000_create_favorites_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('favorites', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->unsignedSmallInteger('anime_id');
+            $table->unsignedBigInteger('anime_id');
             $table->timestamp('created_at')->useCurrent();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
